% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disassembler.R
\name{as.character.bcdf}
\alias{as.character.bcdf}
\title{Convert a bytecode data.frame to a text representation}
\usage{
\method{as.character}{bcdf}(x, incl_expr = FALSE, ...)
}
\arguments{
\item{x}{bytecode data.frame produced by \code{dis()}, \code{disq()} 
and \code{parse_code()}}

\item{incl_expr}{Include the stored expression?  Default: FALSE}

\item{...}{ignored}
}
\description{
Convert a bytecode data.frame to a text representation
}
\examples{
\dontrun{
as.character(disq(1 + 1))
}

}
