% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembler.R
\name{asm}
\alias{asm}
\title{Compile bytecode assembly into an R bytecode object}
\usage{
asm(code, expr_default = quote(stop("[rbytecode]")))
}
\arguments{
\item{code}{R bytecode assembly code as a single string}

\item{expr_default}{The default expression stored with the bytecode.
In general this should be a quoted `call` object, and if you 
are running a standard modern version of R, then it will 
almost never get evaluated/called.
Default: quote(stop("[rbytecode]")).}
}
\value{
bytecode object
}
\description{
Compile bytecode assembly into an R bytecode object
}
\examples{
\dontrun{
code <- "LDCONST 1\nLDCONST 2\nADD\nRETURN"
bc <- asm(code)
eval(bc) # 3
}


}
