% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembler.R
\name{compile_bcdf}
\alias{compile_bcdf}
\title{Compile a bytecode data.frame to an executable R bytecode object}
\usage{
compile_bcdf(bcdf, expr_default = quote(stop("[rbytecode]")))
}
\arguments{
\item{bcdf}{bytecode data.frame}

\item{expr_default}{The default expression stored with the bytecode.
In general this should be a quoted `call` object, and if you 
are running a standard modern version of R, then it will 
almost never get evaluated/called.
Default: quote(stop("[rbytecode]")).}
}
\value{
bytecode object
}
\description{
Compile a bytecode data.frame to an executable R bytecode object
}
\examples{
\dontrun{
code <- "LDCONST 1\nLDCONST 2\nADD\nRETURN"
bcdf <- parse_code(code)
bc   <- compile_bcdf(bcdf)
eval(bc)
}



}
