% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disassembler.R
\name{dis}
\alias{dis}
\title{Disassemble R objects to a bytecode data.frame}
\usage{
dis(x, incl_expr = FALSE, depth = 0L, env = NULL)
}
\arguments{
\item{x}{Possible inputs: language object, call object, compiled bytecode, 
functionquoted expression.}

\item{incl_expr}{include the expression index in the output. Default: FALSE.
This is an advanced feature}

\item{depth}{internal variable tracking recursion depth. Not usually set by user.}

\item{env}{environment for keeping track of labels. Not usually set by user.}
}
\value{
bytecode data.frame (\code{bcdf})
}
\description{
Disassemble R objects to a bytecode data.frame
}
\examples{
\dontrun{
dis(quote(1 + x))
fc <- compiler::cmpfun(\(x) {2 * x + 1}) 
dis(fc)
}

}
