% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disassembler.R
\name{disq}
\alias{disq}
\title{A wrapper for \code{dis()} which accepts unquoted expressions.}
\usage{
disq(x, incl_expr = FALSE)
}
\arguments{
\item{x}{unquoted expression}

\item{incl_expr}{include the expression index in the output. Default: FALSE.
This is an advanced feature}
}
\description{
A wrapper for \code{dis()} which accepts unquoted expressions.
}
\examples{
\dontrun{
disq(1 + 3)
dis(quote(1 + 3))
}

}
