% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_special}
\alias{eval_special}
\title{Evaluate an expression with particular handling for SPECIAL functions}
\usage{
eval_special(bc)
}
\arguments{
\item{bc}{bytecode object to be evaluated}
}
\description{
R has many SPECIAL functions in the base R package that should be called
in bytecode with CALLSPECIAL
}
\details{
But given it's a bit difficult to construct a CALLSPECIAL without just
writing actual R code, there are some SPECIALs which can be handled
by wrapping the special functions in a non-special wrapper.
}
