% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-definitions.R
\docType{data}
\name{ops}
\alias{ops}
\title{Information about all bytecode operators}
\format{
An object of class \code{list} of length 129.
}
\usage{
ops
}
\description{
Information about all bytecode operators
}
\keyword{datasets}
