% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembler.R
\name{parse_code}
\alias{parse_code}
\title{Parse R bytecode assembly into a bytecode data.frame}
\usage{
parse_code(code)
}
\arguments{
\item{code}{Single string containing bytecode instructions}
}
\value{
bytecode data.frame (\code{bcdf})
}
\description{
Parse R bytecode assembly into a bytecode data.frame
}
\examples{
\dontrun{
code <- "LDCONST 1\nLDCONST 2\nADD\nRETURN"
parse_code(code)
}

}
