% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disassembler.R
\name{print.bcdftxt}
\alias{print.bcdftxt}
\title{Print a bytecode data.frame}
\usage{
\method{print}{bcdftxt}(x, ...)
}
\arguments{
\item{x}{Character representation of a bytecode data.frame}

\item{...}{arguments passed to \code{cat()}}
}
\description{
Print a bytecode data.frame
}
\examples{
\dontrun{
bcdf <- disq(1 + x)
txt  <- as.character(bcdf)
txt
}

}
